# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "User service for registration/login of user. It allows to block/unblock other users and update their profiles.",
        "title": "User-Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000/",
    "paths": {
        "/auth/google/": {
            "get": {
                "description": "googleAuth link for getting code (authorization code)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserAuth"
                ],
                "summary": "googleAuth link",
                "operationId": "googleAuth-link",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    }
                }
            }
        },
        "/auth/google/login": {
            "get": {
                "description": "get otp for login of user using google account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserAuth"
                ],
                "summary": "user google login",
                "operationId": "user-google-login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pass the code that we get through the googleAuth link",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.GoogleUserInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    }
                }
            }
        },
...
        "/users/block": {
            "get": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "get list of all blocked users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BlockActions"
                ],
                "summary": "block userList",
                "operationId": "user-blockList",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.BlockUserDetails"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    }
                }
            }
        },
        "/users/block/": {
            "post": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "block another user to hide your details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BlockActions"
                ],
                "summary": "block another user",
                "operationId": "user-block",
                "parameters": [
                    {
                        "description": "The input for user login",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.BlockUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    }
                }
            }
        },
...
        "/users/unblock/{blocked}": {
            "delete": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "unblock another user to show your details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BlockActions"
                ],
                "summary": "unblock another user",
                "operationId": "user-unblock",
                "parameters": [
                    {
                        "type": "string",
                        "description": "blocked",
                        "name": "blocked",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "error.CustomError": {
            "type": "object",
            "required": [
                "errorMessage",
                "statusCode"
            ],
            "properties": {
                "errorMessage": {
                    "type": "string"
                },
                "invalidData": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/error.InvalidData"
                    }
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "error.InvalidData": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "field": {
                    "type": "string"
                }
            }
        },
        "request.BlockUser": {
            "type": "object",
            "required": [
                "blockedUser"
            ],
            "properties": {
                "blockedUser": {
                    "type": "string"
                }
            }
        },
...
        "response.BlockUserDetails": {
            "type": "object",
            "required": [
                "blockedAt",
                "blockedUser",
                "fullname",
                "id"
            ],
            "properties": {
                "blockedAt": {
                    "type": "string"
                },
                "blockedUser": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "response.GoogleUserInfo": {
            "type": "object",
            "required": [
                "email",
                "firstname",
                "lastname"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "UserIDAuth": {
            "type": "apiKey",
            "name": "Auth-user",
            "in": "header"
        },
        "jwtAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}